batch_processing/requirements.txt
google-generativeai
python-dotenv
pillow
